Remove the sleep.
What I have observed is that if I try to set the amperage while the car is still charging, it will fail and I don't have a sleep between stopping the charging and setting the amperage, and that is working fine. So I don't think a sleep between setting the amperage and reading it is needed.