Reset back to minimum when not charging
The charger switches to maximum amperage overnight (when solar produces < 500W). In the morning when solar > 500W but insufficient to start charging, it doesn't do anything. If the car is plugged in during this window, then it will charge at maximum for 5mins.
Instead, make sure to reset amperage back to lowest when there isn't sufficient power to minimize the impact.